+++ CHANGELOG.md
@@ -20,17 +20,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 This tag add all the serialization to save and retrieve photos of a product
 workflow to add these changes are:
 1. Add photo field to the UI ✓
-2. Change POST endpoint to acept multiform and UploadFile
-3. Validation and resizing
+2. Change POST endpoint to acept multiform and UploadFile ✓
+3. Validation and resizing ✓
 4. Upload and get file (locally or S3 bucket)
 5. Pass a photo
 6. Add logic to the waiter template
 
 ### Added
 - Photo field to `src/templates/products.html`
+- `src/api/v1/apps/orders/image_service.py` file
 
 ### Changed
-- N/A
+- POST endpoint for products in `src/api/v1/apps/orders/router.py`
 
 ### Fixed
 - N/A